refactor: rename remote_datasource package to remoteDataSource
