+++ remote_datasource/build.gradle.kts
@@ -50,5 +50,4 @@ dependencies {
     implementation(libs.kotlinx.serialization.json)
     testImplementation(libs.bundles.test.core)
     testImplementation(libs.ktor.client.mock)
-    testImplementation(kotlin("test"))
 }
