+++ remote_datasource/src/main/java/com/baghdad/remoteDataSource/RemoteSearchDataSourceImpl.kt
@@ -1,10 +1,9 @@
-package com.baghdad.remote_datasource
+package com.baghdad.remoteDataSource.util
 
-import com.baghdad.remote_datasource.entity.SearchParameter
-import com.baghdad.remote_datasource.entity.SearchResponse
-import com.baghdad.remote_datasource.entity.toParams
-import com.baghdad.remote_datasource.mapper.toDto
-import com.baghdad.remote_datasource.util.handleRequest
+import com.baghdad.remoteDataSource.util.entity.SearchParameter
+import com.baghdad.remoteDataSource.util.entity.SearchResponse
+import com.baghdad.remoteDataSource.util.entity.toParams
+import com.baghdad.remoteDataSource.util.mapper.toDto
 import com.baghdad.repository.datasource.remote.RemoteSearchDataSource
 import com.baghdad.repository.model.GenreDto
 import com.baghdad.repository.model.SearchResultDto
