+++ remote_datasource/src/main/java/com/baghdad/remoteDataSource/entity/GenreResponse.kt
@@ -1,4 +1,4 @@
-package com.baghdad.remote_datasource.entity
+package com.baghdad.remoteDataSource.util.entity
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
