+++ remote_datasource/src/main/java/com/baghdad/remoteDataSource/mapper/SearchMapper.kt
@@ -1,6 +1,6 @@
-package com.baghdad.remote_datasource.mapper
+package com.baghdad.remoteDataSource.util.mapper
 
-import com.baghdad.remote_datasource.entity.SearchResponse
+import com.baghdad.remoteDataSource.util.entity.SearchResponse
 import com.baghdad.repository.model.ActorDto
 import com.baghdad.repository.model.GenreDto
 import com.baghdad.repository.model.MovieDto
